certa-backend/requirements.txt
absl-py
aiohappyeyeballs
aiosignal
annotated-types
anyio
astunparse
attrs
backoff
bcrypt==3.2.2
beautifulsoup4
blinker
bs4
cachetools
certifi
cffi
charset-normalizer
click
colorama
cssselect
dataclasses-json
distro
dnspython
ecdsa
email_validator
Faker
fastapi
fastapi-cli
feedfinder2
feedparser
filelock
Flask
flatbuffers
frozenlist
gast
glassflow
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-pasta
googleapis-common-protos
greenlet
groq
grpcio
grpcio-status
h11
h2
h5py
hpack
httpcore
httplib2
httptools
httpx
Hypercorn
hyperframe
idna
itsdangerous
jieba3k
Jinja2
jiter
joblib
jsonpath-python
keras
libclang
lxml
lxml_html_clean
Markdown
markdown-it-py
MarkupSafe
marshmallow
mdurl
ml-dtypes
motor
multidict
mypy-extensions
namex
nest-asyncio
newsapi-python
newspaper3k
nltk
numpy
openai
opencv-python
opt_einsum
optree
outcome
packaging
pandas
passlib[bcrypt]==1.7.4
pillow
playwright
playwright-stealth
priority
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydub
pyee
Pygments
pymongo
pyparsing
PySocks
python-dateutil
python-decouple
python-dotenv
python-jose
python-multipart
pytz
PyYAML
regex
requests
requests-file
rich
rich-toolkit
rsa
selenium
setuptools
sgmllib3k
shellingham
six
sniffio
sortedcontainers
soupsieve
SpeechRecognition
starlette
termcolor
tiktoken
tinysegmenter
tldextract
tqdm
trio
trio-websocket
typer
typing-inspect
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
watchfiles
webdriver-manager
websocket-client
websockets
Werkzeug
wheel
wrapt
wsproto
yarl
